Fixes #18. Make CVSTAG work better
The old `CVSTAG.in` constructor code was bad. If a tag exists and it was
older than the current revision hash, the code would still say the tag
was correct.

This is ugly, but seems to work. Someone like @tclune can probably
figure out how to make a CMake macro or function for this code and make
it prettier.